Refactor media exporter code to separate file
